fix: results import
